Added possibility to clip part of the 2D mesh after generating and refining
Added possibility to clip part of the 2D mesh after generating and refining. Before it was only possible to define the parts to be clipped during the initial generating of the mesh. Now the grid can first be refined and clipped after, adding the possibility to have a finer mesh along the clipped parts.

Co-Authored-By: WSRL-HVG <wsrl-hvg@users.noreply.github.com>