Speed up the clean_nodes and clip_nodes
Speed up for the clean nodes function, which looks for edges that do no longer belong to a face. By using contiguous arrays instead of node id tuples, the function is much faster now, and also well applicable to large meshes.

The clip_nodes took a long time to check for edges that have two nodes outside the clip area, but still crossed this area. By not using a geodataframe, and using prepared geometries, the function is faster too know.