Add option for only reading 2d mesh from Ugrid 0.9 file
The function "from_netcdf_old" reads a mesh from a Ugrid 0.9 convention file (with NetNode and NetLink variables). This function is meant for the 2D mesh, it does not distinguish based on NetLinkType. Now the option only2d is added, with which the non 2d part cna be skipped, in case of importing a 1d2d mesh.